scraper/requirements.txt
playwright
redis
python-dotenv
beautifulsoup4
lxml 